chore:主键使用biginteger

+++ database/migrations/2019_11_05_172448_create_mail_table.php
@@ -15,7 +15,7 @@ class CreateMailTable extends Migration
     {
         Schema::create('mail', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('user_id')->default(0);
+            $table->unsignedBigInteger('user_id')->default(0);
             $table->string('guard', 8);
             $table->unsignedInteger('scene_code')->default(0);
             $table->string('title');
